backend/requirements.txt
Flask==2.3.2
flask-pymongo==2.3.0
flask-cors==3.0.10
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
flask-restx==1.3.0
pymongo[srv]==4.4.1
dnspython==2.4.2
Werkzeug==2.3.7
itsdangerous==2.1.2
six==1.16.0
click==8.1.7
MarkupSafe==2.1.3
Jinja2==3.1.2
uvicorn==0.27.0.post1